+++ demo/app/pages/+default/home/home.component.ts
@@ -1,7 +1,7 @@
 import {Component, ChangeDetectionStrategy, Signal, viewChild, effect} from '@angular/core';
 import {ComponentRedirectRoute, ComponentRoute} from '@anglr/common/router';
 import {WithScrollableCssClass} from '@anglr/common';
-import {Select, Option, SelectAbsolute} from '@anglr/select';
+import {Select, Option, SelectAbsolute, SelectEdit} from '@anglr/select';
 import {getValue} from '@anglr/select/extensions';
 
 import {SelectApi} from '../../../../../src/interfaces';
@@ -17,6 +17,7 @@ import {SelectApi} from '../../../../../src/interfaces';
     [
         Select,
         Option,
+        SelectEdit,
         SelectAbsolute,
     ],
     changeDetection: ChangeDetectionStrategy.OnPush