+++ src/directives/index.ts
@@ -3,5 +3,7 @@ export * from './normalStateTagTemplate/normalStateTagTemplate.directive';
 export * from './optionGroupTemplate/optionGroupTemplate.directive';
 export * from './optionTemplate/optionTemplate.directive';
 export * from './selectAbsolute/selectAbsolute.directive';
+export * from './selectEdit/selectEdit.directive';
+export * from './selectFormControl/selectFormControl.directive';
 export * from './selectPlaceholder/selectPlaceholder.directive';
 export * from './withDirectAccess/withDirectAccess.directive';