+++ src/directives/selectEdit/selectEdit.directive.ts
@@ -1,51 +1,41 @@
-// import {Directive, ValueProvider} from '@angular/core';
+import {Directive} from '@angular/core';
+import {RecursivePartial} from '@jscrpt/common';
 
-// import {NORMAL_STATE_TYPE, LIVE_SEARCH_TYPE, POPUP_TYPE, KEYBOARD_HANDLER_TYPE} from '../../misc/types';
-// import {EditNormalStateComponent} from '../../plugins/normalState/edit/editNormalState.component';
-// import {EditLiveSearchComponent} from '../../plugins/liveSearch/edit/editLiveSearch.component';
-// import {EditPopupComponent} from '../../plugins/popup/edit/editPopup.component';
-// import {EditKeyboardHandlerComponent} from '../../plugins/keyboardHandler/components';
-// import {SelectComponent} from '../../components';
-// import {ExcludingOptionsGatherer} from '../../misc/optionsGatherer/types';
+import {provideLiveSearchType, provideNormalStateType} from '../../misc/providers';
+import {EditLiveSearch, EditNormalState} from '../../plugins';
+import {Select} from '../../components/select/select.component';
+import {SelectOptions, SimpleKeyboardHandlerOptions} from '../../interfaces';
 
-// /**
-//  * Directive that applies options for NgSelect which enable usage of NgSelect edit (jira like) style
-//  */
-// @Directive(
-// {
-//     selector: 'ng-select[editStyle]',
-//     providers:
-//     [
-//         <ValueProvider>
-//         {
-//             provide: NORMAL_STATE_TYPE,
-//             useValue: EditNormalStateComponent,
-//         },
-//         <ValueProvider>
-//         {
-//             provide: LIVE_SEARCH_TYPE,
-//             useValue: EditLiveSearchComponent,
-//         },
-//         <ValueProvider>
-//         {
-//             provide: POPUP_TYPE,
-//             useValue: EditPopupComponent,
-//         },
-//         <ValueProvider>
-//         {
-//             provide: KEYBOARD_HANDLER_TYPE,
-//             useValue: EditKeyboardHandlerComponent,
-//         },
-//     ],
-// })
-// export class SelectEditDirective
-// {
-//     //######################### constructor #########################
-//     constructor(select: SelectComponent)
-//     {
-//         select.selectOptions =
-//         {
-//             optionsGatherer: new ExcludingOptionsGatherer(),
-//         };
-//     }
-// }
+/**
+ * Directive that applies options for Select which switch it to edit mode.
+ */
+@Directive(
+{
+    selector: 'ng-select[edit]',
+    providers:
+    [
+        provideNormalStateType(EditNormalState),
+        provideLiveSearchType(EditLiveSearch),
+    ],
+})
+export class SelectEdit<TValue = unknown>
+{
+    //######################### constructor #########################
+    constructor(select: Select<TValue>)
+    {
+        select.selectOptions =
+        {
+            plugins:
+            {
+                keyboardHandler:
+                {
+                    options: <RecursivePartial<SimpleKeyboardHandlerOptions>>
+                    {
+                        alphanumericSearch: false,
+                        liveSearchEvents: true,
+                    },
+                },
+            },
+        } as SelectOptions<TValue>;
+    }
+}