+++ src/directives/selectPlaceholder/selectPlaceholder.directive.ts
@@ -1,7 +1,7 @@
 import {Directive, effect, input, InputSignal} from '@angular/core';
 
-import {Select} from '../../components';
 import {SelectOptions} from '../../interfaces';
+import {Select} from '../../components/select/select.component';
 
 /**
  * Directive used for setting live search placeholder text